Restructure web crate with more modules
