+++ rust/maprando-web/src/main.rs
@@ -1,1932 +1,31 @@
-mod logic;
-mod m_web;
+mod logic_helper;
+mod web;
 
-use std::path::Path;
-use std::time::SystemTime;
-
-use crate::m_web::{AppData, PresetData, VersionInfo, HQ_VIDEO_URL_ROOT};
-use actix_easy_multipart::bytes::Bytes;
-use actix_easy_multipart::text::Text;
-use actix_easy_multipart::{MultipartForm, MultipartFormConfig};
-use actix_web::http::header::{
-    self, CacheControl, CacheDirective, ContentDisposition, DispositionParam, DispositionType,
+use crate::{
+    logic_helper::LogicData,
+    web::{AppData, PresetData, VersionInfo, HQ_VIDEO_URL_ROOT, VERSION},
 };
-use actix_web::middleware::Compress;
-use actix_web::middleware::Logger;
-use actix_web::{get, post, web, App, HttpRequest, HttpResponse, HttpServer, Responder};
-use anyhow::{bail, Context, Result};
-use askama::Template;
+use actix_easy_multipart::MultipartFormConfig;
+use actix_web::{
+    middleware::{Compress, Logger},
+    App, HttpServer,
+};
 use clap::Parser;
-use hashbrown::{HashMap, HashSet};
-use log::{error, info};
-use maprando::customize::mosaic::MosaicTheme;
-use maprando::customize::samus_sprite::SamusSpriteCategory;
-use maprando::customize::{
-    customize_rom, parse_controller_button, ControllerButton, ControllerConfig, CustomizeSettings,
-    DoorTheme, MusicSettings, PaletteTheme, ShakingSetting, TileTheme,
+use hashbrown::HashSet;
+use log::info;
+use maprando::{
+    customize::{mosaic::MosaicTheme, samus_sprite::SamusSpriteCategory},
+    game_data::GameData,
+    map_repository::MapRepository,
+    preset::Preset,
+    seed_repository::SeedRepository,
 };
-use maprando::game_data::{Capacity, GameData, IndexedVec, Item, LinksDataGroup, Map};
-use maprando::map_repository::MapRepository;
-use maprando::patch::ips_write::create_ips_patch;
-use maprando::patch::{make_rom, Rom};
-use maprando::preset::Preset;
-use maprando::randomize::{
-    filter_links, randomize_doors, randomize_map_areas, AreaAssignment, DebugOptions,
-    DifficultyConfig, DoorLocksSize, DoorsMode, EtankRefill, ItemDotChange, ItemMarkers,
-    ItemPlacementStyle, ItemPriorityGroup, ItemPriorityStrength, MotherBrainFight, Objective,
-    Randomization, Randomizer, SaveAnimals, StartLocationMode, WallJump,
-};
-use maprando::seed_repository::{Seed, SeedFile, SeedRepository};
-use maprando::spoiler_map;
-use rand::{RngCore, SeedableRng};
-use serde_derive::{Deserialize, Serialize};
-use std::time::Instant;
-
-use crate::logic::LogicData;
-use crate::m_web::VERSION;
+use std::{path::Path, time::Instant};
+use web::{about, generate, home, logic, randomize, releases, seed};
 
 const VISUALIZER_PATH: &'static str = "../visualizer/";
 const TECH_GIF_PATH: &'static str = "static/tech_gifs/";
 const NOTABLE_GIF_PATH: &'static str = "static/notable_gifs/";
-
-#[derive(Template)]
-#[template(path = "errors/missing_input_rom.html")]
-struct MissingInputRomTemplate {}
-
-#[derive(Template)]
-#[template(path = "errors/invalid_rom.html")]
-struct InvalidRomTemplate {}
-
-#[derive(Template)]
-#[template(path = "errors/seed_not_found.html")]
-struct SeedNotFoundTemplate {}
-
-#[derive(Template)]
-#[template(path = "errors/room_not_found.html")]
-struct RoomNotFoundTemplate {}
-
-#[derive(Template)]
-#[template(path = "errors/file_not_found.html")]
-struct FileNotFoundTemplate {}
-
-#[derive(Template)]
-#[template(path = "errors/invalid_token.html")]
-struct InvalidTokenTemplate {}
-
-#[derive(Template)]
-#[template(path = "errors/already_unlocked.html")]
-struct AlreadyUnlockedTemplate {}
-
-#[derive(Template)]
-#[template(path = "home.html")]
-struct HomeTemplate {
-    version_info: VersionInfo,
-}
-
-#[derive(Template)]
-#[template(path = "releases.html")]
-struct ReleasesTemplate {
-    version_info: VersionInfo,
-}
-
-#[derive(Template)]
-#[template(path = "about.html")]
-struct AboutTemplate {
-    version_info: VersionInfo,
-    sprite_artists: Vec<String>,
-}
-
-impl AboutTemplate {
-    fn sprite_artists(&self) -> String {
-        self.sprite_artists
-            .iter()
-            .map(|x| format!("<i>{}</i>", x))
-            .collect::<Vec<String>>()
-            .join(", ")
-    }
-}
-
-#[derive(Template)]
-#[template(path = "generate/main.html")]
-struct GenerateTemplate<'a> {
-    version_info: VersionInfo,
-    progression_rates: Vec<&'static str>,
-    item_placement_styles: Vec<&'static str>,
-    objectives: Vec<&'static str>,
-    preset_data: &'a [PresetData],
-    item_priorities: Vec<String>,
-    item_pool_multiple: Vec<String>,
-    starting_items_multiple: Vec<String>,
-    starting_items_single: Vec<String>,
-    prioritizable_items: Vec<String>,
-    tech_description: &'a HashMap<String, String>,
-    tech_dependencies: &'a HashMap<String, Vec<String>>,
-    strat_dependencies: &'a HashMap<String, Vec<String>>,
-    _strat_area: &'a HashMap<String, String>,
-    strat_description: &'a HashMap<String, String>,
-    strat_id_by_name: &'a HashMap<String, usize>,
-    tech_gif_listing: &'a HashSet<String>,
-    notable_gif_listing: &'a HashSet<String>,
-    tech_strat_counts: &'a HashMap<String, usize>,
-    hq_video_url_root: &'a str,
-}
-
-#[get("/")]
-async fn home(app_data: web::Data<AppData>) -> impl Responder {
-    let home_template = HomeTemplate {
-        version_info: app_data.version_info.clone(),
-    };
-    HttpResponse::Ok().body(home_template.render().unwrap())
-}
-
-#[get("/releases")]
-async fn releases(app_data: web::Data<AppData>) -> impl Responder {
-    let changes_template = ReleasesTemplate {
-        version_info: app_data.version_info.clone(),
-    };
-    HttpResponse::Ok().body(changes_template.render().unwrap())
-}
-
-#[get("/about")]
-async fn about(app_data: web::Data<AppData>) -> impl Responder {
-    let mut sprite_artists = vec![];
-
-    for category in &app_data.samus_sprite_categories {
-        for info in &category.sprites {
-            for author in &info.authors {
-                if info.display_name != "Samus" {
-                    sprite_artists.push(author.clone());
-                }
-            }
-        }
-    }
-    sprite_artists.sort_by_key(|x| x.to_lowercase());
-    sprite_artists.dedup();
-    let about_template = AboutTemplate {
-        version_info: app_data.version_info.clone(),
-        sprite_artists,
-    };
-    HttpResponse::Ok().body(about_template.render().unwrap())
-}
-
-#[get("/generate")]
-async fn generate(app_data: web::Data<AppData>) -> impl Responder {
-    let item_pool_multiple: Vec<String> = ["Missile", "ETank", "ReserveTank", "Super", "PowerBomb"]
-        .into_iter()
-        .map(|x| x.to_string())
-        .collect();
-
-    let starting_items_multiple: Vec<String> =
-        ["Missile", "ETank", "ReserveTank", "Super", "PowerBomb"]
-            .into_iter()
-            .map(|x| x.to_string())
-            .collect();
-
-    let starting_items_single: Vec<String> = [
-        "Charge",
-        "Ice",
-        "Wave",
-        "Spazer",
-        "Plasma",
-        "XRayScope",
-        "Morph",
-        "Bombs",
-        "Grapple",
-        "HiJump",
-        "SpeedBooster",
-        "SpringBall",
-        "SpaceJump",
-        "ScrewAttack",
-        "Varia",
-        "Gravity",
-    ]
-    .into_iter()
-    .map(|x| x.to_string())
-    .collect();
-
-    let prioritizable_items: Vec<String> = [
-        "ETank",
-        "ReserveTank",
-        "Super",
-        "PowerBomb",
-        "Charge",
-        "Ice",
-        "Wave",
-        "Spazer",
-        "Plasma",
-        "XRayScope",
-        "Morph",
-        "Bombs",
-        "Grapple",
-        "HiJump",
-        "SpeedBooster",
-        "SpringBall",
-        "WallJump",
-        "SpaceJump",
-        "ScrewAttack",
-        "Varia",
-        "Gravity",
-    ]
-    .into_iter()
-    .map(|x| x.to_string())
-    .collect();
-    let generate_template = GenerateTemplate {
-        version_info: app_data.version_info.clone(),
-        progression_rates: vec!["Fast", "Uniform", "Slow"],
-        item_placement_styles: vec!["Neutral", "Forced"],
-        objectives: vec![
-            "None",
-            "Bosses",
-            "Minibosses",
-            "Metroids",
-            "Chozos",
-            "Pirates",
-            "Random",
-        ],
-        item_pool_multiple,
-        starting_items_multiple,
-        starting_items_single,
-        item_priorities: vec!["Early", "Default", "Late"]
-            .iter()
-            .map(|x| x.to_string())
-            .collect(),
-        prioritizable_items,
-        preset_data: &app_data.preset_data,
-        tech_description: &app_data.game_data.tech_description,
-        tech_dependencies: &app_data.game_data.tech_dependencies,
-        strat_dependencies: &app_data.game_data.strat_dependencies,
-        _strat_area: &app_data.game_data.strat_area,
-        strat_description: &app_data.game_data.strat_description,
-        strat_id_by_name: &app_data.game_data.notable_strat_isv.index_by_key,
-        tech_gif_listing: &app_data.tech_gif_listing,
-        notable_gif_listing: &app_data.notable_gif_listing,
-        tech_strat_counts: &app_data.logic_data.tech_strat_counts,
-        hq_video_url_root: HQ_VIDEO_URL_ROOT,
-    };
-    HttpResponse::Ok().body(generate_template.render().unwrap())
-}
-
-#[derive(MultipartForm)]
-struct RandomizeRequest {
-    rom: Bytes,
-    preset: Option<Text<String>>,
-    shinespark_tiles: Text<f32>,
-    heated_shinespark_tiles: Text<f32>,
-    speed_ball_tiles: Text<f32>,
-    shinecharge_leniency_frames: Text<Capacity>,
-    resource_multiplier: Text<f32>,
-    gate_glitch_leniency: Text<Capacity>,
-    door_stuck_leniency: Text<Capacity>,
-    phantoon_proficiency: Text<f32>,
-    draygon_proficiency: Text<f32>,
-    ridley_proficiency: Text<f32>,
-    botwoon_proficiency: Text<f32>,
-    mother_brain_proficiency: Text<f32>,
-    escape_timer_multiplier: Text<f32>,
-    tech_json: Text<String>,
-    strat_json: Text<String>,
-    progression_rate: Text<String>,
-    item_placement_style: Text<String>,
-    item_priority_strength: Text<String>,
-    random_tank: Text<String>,
-    spazer_before_plasma: Text<String>,
-    stop_item_placement_early: Text<String>,
-    item_progression_preset: Option<Text<String>>,
-    item_pool_json: Text<String>,
-    starting_item_json: Text<String>,
-    item_priority_json: Text<String>,
-    filler_items_json: Text<String>,
-    race_mode: Text<String>,
-    random_seed: Text<String>,
-    spoiler_token: Text<String>,
-    quality_of_life_preset: Option<Text<String>>,
-    supers_double: Text<bool>,
-    mother_brain_fight: Text<String>,
-    escape_enemies_cleared: Text<bool>,
-    escape_refill: Text<bool>,
-    escape_movement_items: Text<bool>,
-    mark_map_stations: Text<bool>,
-    room_outline_revealed: Text<bool>,
-    transition_letters: Text<bool>,
-    door_locks_size: Text<String>,
-    item_markers: Text<String>,
-    item_dot_change: Text<String>,
-    all_items_spawn: Text<bool>,
-    acid_chozo: Text<bool>,
-    buffed_drops: Text<bool>,
-    fast_elevators: Text<bool>,
-    fast_doors: Text<bool>,
-    fast_pause_menu: Text<bool>,
-    respin: Text<bool>,
-    infinite_space_jump: Text<bool>,
-    momentum_conservation: Text<bool>,
-    objectives: Text<String>,
-    doors: Text<String>,
-    start_location: Text<String>,
-    map_layout: Text<String>,
-    save_animals: Text<String>,
-    early_save: Text<bool>,
-    area_assignment: Text<String>,
-    wall_jump: Text<String>,
-    etank_refill: Text<String>,
-    maps_revealed: Text<String>,
-    map_station_reveal: Text<String>,
-    energy_free_shinesparks: Text<bool>,
-    ultra_low_qol: Text<bool>,
-}
-
-#[derive(Deserialize)]
-struct UnlockRequest {
-    spoiler_token: String,
-}
-
-#[derive(MultipartForm)]
-struct CustomizeRequest {
-    rom: Bytes,
-    samus_sprite: Text<String>,
-    etank_color: Text<String>,
-    reserve_hud_style: Text<bool>,
-    room_palettes: Text<String>,
-    tile_theme: Text<String>,
-    door_theme: Text<String>,
-    music: Text<String>,
-    disable_beeping: Text<bool>,
-    shaking: Text<String>,
-    vanilla_screw_attack_animation: Text<bool>,
-    control_shot: Text<String>,
-    control_jump: Text<String>,
-    control_dash: Text<String>,
-    control_item_select: Text<String>,
-    control_item_cancel: Text<String>,
-    control_angle_up: Text<String>,
-    control_angle_down: Text<String>,
-    spin_lock_left: Option<Text<String>>,
-    spin_lock_right: Option<Text<String>>,
-    spin_lock_up: Option<Text<String>>,
-    spin_lock_down: Option<Text<String>>,
-    spin_lock_x: Option<Text<String>>,
-    spin_lock_y: Option<Text<String>>,
-    spin_lock_a: Option<Text<String>>,
-    spin_lock_b: Option<Text<String>>,
-    spin_lock_l: Option<Text<String>>,
-    spin_lock_r: Option<Text<String>>,
-    spin_lock_select: Option<Text<String>>,
-    spin_lock_start: Option<Text<String>>,
-    quick_reload_left: Option<Text<String>>,
-    quick_reload_right: Option<Text<String>>,
-    quick_reload_up: Option<Text<String>>,
-    quick_reload_down: Option<Text<String>>,
-    quick_reload_x: Option<Text<String>>,
-    quick_reload_y: Option<Text<String>>,
-    quick_reload_a: Option<Text<String>>,
-    quick_reload_b: Option<Text<String>>,
-    quick_reload_l: Option<Text<String>>,
-    quick_reload_r: Option<Text<String>>,
-    quick_reload_select: Option<Text<String>>,
-    quick_reload_start: Option<Text<String>>,
-    moonwalk: Text<bool>,
-}
-
-#[derive(Serialize, Deserialize)]
-struct SeedData {
-    version: usize,
-    timestamp: usize,
-    peer_addr: String,
-    http_headers: serde_json::Map<String, serde_json::Value>,
-    random_seed: usize,
-    map_seed: usize,
-    door_randomization_seed: usize,
-    item_placement_seed: usize,
-    race_mode: bool,
-    preset: Option<String>,
-    item_progression_preset: Option<String>,
-    difficulty: DifficultyConfig,
-    quality_of_life_preset: Option<String>,
-    supers_double: bool,
-    mother_brain_fight: String,
-    escape_enemies_cleared: bool,
-    escape_refill: bool,
-    escape_movement_items: bool,
-    mark_map_stations: bool,
-    transition_letters: bool,
-    item_markers: String,
-    item_dot_change: String,
-    all_items_spawn: bool,
-    acid_chozo: bool,
-    buffed_drops: bool,
-    fast_elevators: bool,
-    fast_doors: bool,
-    fast_pause_menu: bool,
-    respin: bool,
-    infinite_space_jump: bool,
-    momentum_conservation: bool,
-    objectives: String,
-    doors: String,
-    start_location_mode: String,
-    map_layout: String,
-    save_animals: String,
-    early_save: bool,
-    area_assignment: String,
-    wall_jump: String,
-    etank_refill: String,
-    maps_revealed: String,
-    vanilla_map: bool,
-    ultra_low_qol: bool,
-}
-
-#[derive(Template)]
-#[template(path = "seed/seed_header.html")]
-struct SeedHeaderTemplate<'a> {
-    seed_name: String,
-    timestamp: usize, // Milliseconds since UNIX epoch
-    random_seed: usize,
-    version_info: VersionInfo,
-    race_mode: bool,
-    preset: String,
-    item_progression_preset: String,
-    progression_rate: String,
-    random_tank: bool,
-    filler_items: Vec<String>,
-    semi_filler_items: Vec<String>,
-    early_filler_items: Vec<String>,
-    item_placement_style: String,
-    difficulty: &'a DifficultyConfig,
-    _notable_strats: Vec<String>,
-    quality_of_life_preset: String,
-    supers_double: bool,
-    mother_brain_fight: String,
-    escape_enemies_cleared: bool,
-    escape_refill: bool,
-    escape_movement_items: bool,
-    mark_map_stations: bool,
-    transition_letters: bool,
-    item_markers: String,
-    item_dot_change: String,
-    all_items_spawn: bool,
-    acid_chozo: bool,
-    buffed_drops: bool,
-    fast_elevators: bool,
-    fast_doors: bool,
-    fast_pause_menu: bool,
-    respin: bool,
-    infinite_space_jump: bool,
-    momentum_conservation: bool,
-    objectives: String,
-    doors: String,
-    start_location_mode: String,
-    map_layout: String,
-    save_animals: String,
-    early_save: bool,
-    area_assignment: String,
-    ultra_low_qol: bool,
-    preset_data: &'a [PresetData],
-    enabled_tech: HashSet<String>,
-    enabled_notables: HashSet<String>,
-}
-
-impl<'a> SeedHeaderTemplate<'a> {
-    fn percent_enabled(&self, p: &PresetData) -> isize {
-        let tech_enabled_count = p
-            .preset
-            .tech
-            .iter()
-            .filter(|&x| self.enabled_tech.contains(x))
-            .count();
-        let notable_enabled_count = p
-            .preset
-            .notable_strats
-            .iter()
-            .filter(|&x| self.enabled_notables.contains(x))
-            .count();
-        let total_enabled_count = tech_enabled_count + notable_enabled_count;
-        let total_count = p.preset.tech.len() + p.preset.notable_strats.len();
-        let frac_enabled = (total_enabled_count as f32) / (total_count as f32);
-        let mut percent_enabled = (frac_enabled * 100.0) as isize;
-        if percent_enabled == 0 && frac_enabled > 0.0 {
-            percent_enabled = 1;
-        }
-        if percent_enabled == 100 && frac_enabled < 1.0 {
-            percent_enabled = 99;
-        }
-        percent_enabled
-    }
-
-    fn item_pool_strs(&self) -> String {
-        self.difficulty
-            .item_pool
-            .iter()
-            .map(|(x, cnt)| {
-                if *cnt > 1 {
-                    format!("{:?} ({})", x, cnt)
-                } else {
-                    format!("{:?}", x)
-                }
-            })
-            .collect::<Vec<String>>()
-            .join(", ")
-    }
-
-    fn starting_items_strs(&self) -> String {
-        self.difficulty
-            .starting_items
-            .iter()
-            .map(|(x, cnt)| {
-                if *cnt > 1 {
-                    format!("{:?} ({})", x, cnt)
-                } else {
-                    format!("{:?}", x)
-                }
-            })
-            .collect::<Vec<String>>()
-            .join(", ")
-    }
-}
-
-impl<'a> SeedHeaderTemplate<'a> {
-    fn game_variations(&self) -> Vec<&str> {
-        let mut game_variations = vec![];
-        if self.area_assignment == "Random" {
-            game_variations.push("Random area assignment");
-        }
-        if self.item_dot_change == "Disappear" {
-            game_variations.push("Item dots disappear after collection");
-        }
-        if !self.transition_letters {
-            game_variations.push("Area transitions marked as arrows");
-        }
-        if self.difficulty.door_locks_size == DoorLocksSize::Small {
-            game_variations.push("Door locks drawn smaller on map");
-        }
-        match self.difficulty.wall_jump {
-            WallJump::Collectible => {
-                game_variations.push("Collectible wall jump");
-            }
-            _ => {}
-        }
-        match self.difficulty.etank_refill {
-            EtankRefill::Disabled => {
-                game_variations.push("E-Tank refill disabled");
-            }
-            EtankRefill::Full => {
-                game_variations.push("E-Tanks refill reserves");
-            }
-            _ => {}
-        }
-        if self.difficulty.maps_revealed == maprando::randomize::MapsRevealed::Partial {
-            game_variations.push("Maps partially revealed from start");
-        }
-        if self.difficulty.maps_revealed == maprando::randomize::MapsRevealed::Full {
-            game_variations.push("Maps revealed from start");
-        }
-        if self.difficulty.map_station_reveal == maprando::randomize::MapStationReveal::Partial {
-            game_variations.push("Map stations give partial reveal");
-        }
-
-        if self.difficulty.energy_free_shinesparks {
-            game_variations.push("Energy-free shinesparks");
-        }
-        if self.ultra_low_qol {
-            game_variations.push("Ultra-low quality of life");
-        }
-        game_variations
-    }
-}
-
-#[derive(Template)]
-#[template(path = "seed/seed_footer.html")]
-struct SeedFooterTemplate {
-    race_mode: bool,
-    all_items_spawn: bool,
-    supers_double: bool,
-    ultra_low_qol: bool,
-}
-
-#[derive(Template)]
-#[template(path = "seed/customize_seed.html")]
-struct CustomizeSeedTemplate {
-    version_info: VersionInfo,
-    spoiler_token_prefix: String,
-    unlocked_timestamp_str: String,
-    seed_header: String,
-    seed_footer: String,
-    samus_sprite_categories: Vec<SamusSpriteCategory>,
-    etank_colors: Vec<Vec<String>>,
-    mosaic_themes: Vec<MosaicTheme>,
-}
-
-fn render_seed(
-    seed_name: &str,
-    seed_data: &SeedData,
-    app_data: &AppData,
-) -> Result<(String, String)> {
-    let notable_strats: Vec<String> = seed_data
-        .difficulty
-        .notable_strats
-        .iter()
-        .cloned()
-        .collect();
-    let enabled_tech: HashSet<String> = seed_data.difficulty.tech.iter().cloned().collect();
-    let enabled_notables: HashSet<String> = seed_data
-        .difficulty
-        .notable_strats
-        .iter()
-        .cloned()
-        .collect();
-    let seed_header_template = SeedHeaderTemplate {
-        seed_name: seed_name.to_string(),
-        version_info: app_data.version_info.clone(),
-        random_seed: seed_data.random_seed,
-        race_mode: seed_data.race_mode,
-        timestamp: seed_data.timestamp,
-        preset: seed_data.preset.clone().unwrap_or("Custom".to_string()),
-        item_progression_preset: seed_data
-            .item_progression_preset
-            .clone()
-            .unwrap_or("Custom".to_string()),
-        progression_rate: format!("{:?}", seed_data.difficulty.progression_rate),
-        random_tank: seed_data.difficulty.random_tank,
-        filler_items: seed_data
-            .difficulty
-            .filler_items
-            .iter()
-            .filter(|&&x| x != Item::Nothing)
-            .map(|x| format!("{:?}", x))
-            .collect(),
-        semi_filler_items: seed_data
-            .difficulty
-            .semi_filler_items
-            .iter()
-            .map(|x| format!("{:?}", x))
-            .collect(),
-        early_filler_items: seed_data
-            .difficulty
-            .early_filler_items
-            .iter()
-            .map(|x| format!("{:?}", x))
-            .collect(),
-        item_placement_style: format!("{:?}", seed_data.difficulty.item_placement_style),
-        difficulty: &seed_data.difficulty,
-        _notable_strats: notable_strats,
-        quality_of_life_preset: seed_data
-            .quality_of_life_preset
-            .clone()
-            .unwrap_or("Custom".to_string()),
-        supers_double: seed_data.supers_double,
-        mother_brain_fight: seed_data.mother_brain_fight.clone(),
-        escape_enemies_cleared: seed_data.escape_enemies_cleared,
-        escape_refill: seed_data.escape_refill,
-        escape_movement_items: seed_data.escape_movement_items,
-        mark_map_stations: seed_data.mark_map_stations,
-        item_markers: seed_data.item_markers.clone(),
-        item_dot_change: seed_data.item_dot_change.clone(),
-        transition_letters: seed_data.transition_letters,
-        all_items_spawn: seed_data.all_items_spawn,
-        acid_chozo: seed_data.acid_chozo,
-        buffed_drops: seed_data.buffed_drops,
-        fast_elevators: seed_data.fast_elevators,
-        fast_doors: seed_data.fast_doors,
-        fast_pause_menu: seed_data.fast_pause_menu,
-        respin: seed_data.respin,
-        infinite_space_jump: seed_data.infinite_space_jump,
-        momentum_conservation: seed_data.momentum_conservation,
-        objectives: seed_data.objectives.clone(),
-        doors: seed_data.doors.clone(),
-        start_location_mode: seed_data.start_location_mode.clone(),
-        map_layout: seed_data.map_layout.clone(),
-        save_animals: seed_data.save_animals.clone(),
-        early_save: seed_data.early_save,
-        area_assignment: seed_data.area_assignment.clone(),
-        ultra_low_qol: seed_data.ultra_low_qol,
-        preset_data: &app_data.preset_data,
-        enabled_tech,
-        enabled_notables,
-    };
-    let seed_header_html = seed_header_template.render()?;
-
-    let seed_footer_template = SeedFooterTemplate {
-        race_mode: seed_data.race_mode,
-        all_items_spawn: seed_data.all_items_spawn,
-        supers_double: seed_data.supers_double,
-        ultra_low_qol: seed_data.ultra_low_qol,
-    };
-    let seed_footer_html = seed_footer_template.render()?;
-    Ok((seed_header_html, seed_footer_html))
-}
-
-async fn check_seed_exists(seed_name: &str, app_data: &AppData) -> bool {
-    app_data
-        .seed_repository
-        .get_file(seed_name, "seed_data.json")
-        .await
-        .is_ok()
-}
-
-async fn save_seed(
-    seed_name: &str,
-    seed_data: &SeedData,
-    spoiler_token: &str,
-    vanilla_rom: &Rom,
-    output_rom: &Rom,
-    randomization: &Randomization,
-    app_data: &AppData,
-) -> Result<()> {
-    if check_seed_exists(seed_name, app_data).await {
-        bail!("Seed name already exists: {}", seed_name);
-    }
-
-    let mut files: Vec<SeedFile> = Vec::new();
-
-    // Write the seed data JSON. This contains details about the seed and request origin,
-    // so to protect user privacy and the integrity of race ROMs we do not make it public.
-    let seed_data_str = serde_json::to_vec_pretty(&seed_data).unwrap();
-    files.push(SeedFile::new("seed_data.json", seed_data_str.to_vec()));
-
-    // Write the ROM patch.
-    let patch_ips = create_ips_patch(&vanilla_rom.data, &output_rom.data);
-    files.push(SeedFile::new("patch.ips", patch_ips));
-
-    // Write the seed header HTML and footer HTML
-    let (seed_header_html, seed_footer_html) = render_seed(seed_name, seed_data, app_data)?;
-    files.push(SeedFile::new(
-        "seed_header.html",
-        seed_header_html.into_bytes(),
-    ));
-    files.push(SeedFile::new(
-        "seed_footer.html",
-        seed_footer_html.into_bytes(),
-    ));
-
-    let prefix = if seed_data.race_mode {
-        "locked"
-    } else {
-        "public"
-    };
-
-    if seed_data.race_mode {
-        files.push(SeedFile::new(
-            "spoiler_token.txt",
-            spoiler_token.as_bytes().to_vec(),
-        ));
-    }
-
-    // Write the map data
-    files.push(SeedFile::new(
-        "map.json",
-        serde_json::to_string(&randomization.map)?
-            .as_bytes()
-            .to_vec(),
-    ));
-
-    // Write the spoiler log
-    let spoiler_bytes = serde_json::to_vec_pretty(&randomization.spoiler_log).unwrap();
-    files.push(SeedFile::new(
-        &format!("{}/spoiler.json", prefix),
-        spoiler_bytes,
-    ));
-
-    // Write the spoiler maps
-    let spoiler_maps =
-        spoiler_map::get_spoiler_map(&output_rom, &randomization.map, &app_data.game_data).unwrap();
-    files.push(SeedFile::new(
-        &format!("{}/map-assigned.png", prefix),
-        spoiler_maps.assigned,
-    ));
-    files.push(SeedFile::new(
-        &format!("{}/map-vanilla.png", prefix),
-        spoiler_maps.vanilla,
-    ));
-    files.push(SeedFile::new(
-        &format!("{}/map-grid.png", prefix),
-        spoiler_maps.grid,
-    ));
-
-    // Write the spoiler visualizer
-    for (filename, data) in &app_data.visualizer_files {
-        let path = format!("{}/visualizer/{}", prefix, filename);
-        files.push(SeedFile::new(&path, data.clone()));
-    }
-
-    let seed = Seed {
-        name: seed_name.to_string(),
-        files,
-    };
-    app_data.seed_repository.put_seed(seed).await?;
-    Ok(())
-}
-
-#[get("/seed/{name}")]
-async fn view_seed_redirect(info: web::Path<(String,)>) -> impl Responder {
-    // Redirect to the seed page (with the trailing slash):
-    HttpResponse::Found()
-        .insert_header((header::LOCATION, format!("{}/", info.0)))
-        .finish()
-}
-
-#[get("/logic")]
-async fn logic_main(app_data: web::Data<AppData>) -> impl Responder {
-    HttpResponse::Ok().body(app_data.logic_data.index_html.clone())
-}
-
-#[get("/logic/room/{name}")]
-async fn logic_room(info: web::Path<(String,)>, app_data: web::Data<AppData>) -> impl Responder {
-    let room_name = &info.0;
-    if let Some(html) = app_data.logic_data.room_html.get(room_name) {
-        HttpResponse::Ok().body(html.clone())
-    } else {
-        let template = RoomNotFoundTemplate {};
-        HttpResponse::NotFound().body(template.render().unwrap())
-    }
-}
-
-#[get("/logic/room/{room_name}/{from_node}/{to_node}/{strat_name}")]
-async fn logic_strat(
-    info: web::Path<(String, usize, usize, String)>,
-    app_data: web::Data<AppData>,
-) -> impl Responder {
-    let room_name = &info.0;
-    let from_node = info.1;
-    let to_node = info.2;
-    let strat_name = &info.3;
-    if let Some(html) = app_data.logic_data.strat_html.get(&(
-        room_name.clone(),
-        from_node,
-        to_node,
-        strat_name.clone(),
-    )) {
-        HttpResponse::Ok().body(html.clone())
-    } else {
-        let template = RoomNotFoundTemplate {};
-        HttpResponse::NotFound().body(template.render().unwrap())
-    }
-}
-
-#[get("/logic/tech/{name}")]
-async fn logic_tech(info: web::Path<(String,)>, app_data: web::Data<AppData>) -> impl Responder {
-    let tech_name = &info.0;
-    if let Some(html) = app_data.logic_data.tech_html.get(tech_name) {
-        HttpResponse::Ok().body(html.clone())
-    } else {
-        let template = RoomNotFoundTemplate {};
-        HttpResponse::NotFound().body(template.render().unwrap())
-    }
-}
-
-#[get("/seed/{name}/")]
-async fn view_seed(info: web::Path<(String,)>, app_data: web::Data<AppData>) -> impl Responder {
-    let seed_name = &info.0;
-    let (seed_header, seed_footer, unlocked_timestamp_str, spoiler_token) = futures::join!(
-        app_data
-            .seed_repository
-            .get_file(seed_name, "seed_header.html"),
-        app_data
-            .seed_repository
-            .get_file(seed_name, "seed_footer.html"),
-        app_data
-            .seed_repository
-            .get_file(seed_name, "unlocked_timestamp.txt"),
-        app_data
-            .seed_repository
-            .get_file(seed_name, "spoiler_token.txt"),
-    );
-    let spoiler_token = String::from_utf8(spoiler_token.unwrap_or(vec![])).unwrap();
-    let spoiler_token_prefix = if spoiler_token.is_empty() {
-        "".to_string()
-    } else {
-        spoiler_token[0..16].to_string()
-    };
-    match (seed_header, seed_footer) {
-        (Ok(header), Ok(footer)) => {
-            let customize_template = CustomizeSeedTemplate {
-                version_info: app_data.version_info.clone(),
-                unlocked_timestamp_str: String::from_utf8(unlocked_timestamp_str.unwrap_or(vec![]))
-                    .unwrap(),
-                spoiler_token_prefix: spoiler_token_prefix.to_string(),
-                seed_header: String::from_utf8(header.to_vec()).unwrap(),
-                seed_footer: String::from_utf8(footer.to_vec()).unwrap(),
-                samus_sprite_categories: app_data.samus_sprite_categories.clone(),
-                etank_colors: app_data.etank_colors.clone(),
-                mosaic_themes: app_data.mosaic_themes.clone(),
-            };
-            // We use a no-cache directive to prevent problems when we change the JavaScript.
-            // Probably better would be to properly version the JS and control cache that way.
-            HttpResponse::Ok()
-                .insert_header(CacheControl(vec![CacheDirective::NoCache]))
-                .body(customize_template.render().unwrap())
-        }
-        (Err(err), _) => {
-            error!("{}", err.to_string());
-            let template = SeedNotFoundTemplate {};
-            HttpResponse::NotFound().body(template.render().unwrap())
-        }
-        (_, Err(err)) => {
-            error!("{}", err.to_string());
-            let template = SeedNotFoundTemplate {};
-            HttpResponse::NotFound().body(template.render().unwrap())
-        }
-    }
-}
-
-#[post("/seed/{name}/unlock")]
-async fn unlock_seed(
-    req: web::Form<UnlockRequest>,
-    info: web::Path<(String,)>,
-    app_data: web::Data<AppData>,
-) -> impl Responder {
-    let seed_name = &info.0;
-    let seed_spoiler_token = app_data
-        .seed_repository
-        .get_file(seed_name, "spoiler_token.txt")
-        .await
-        .unwrap();
-
-    if req.spoiler_token.as_bytes() == seed_spoiler_token {
-        let unlocked_timestamp_data = app_data
-            .seed_repository
-            .get_file(seed_name, "unlocked_timestamp.txt")
-            .await;
-        if unlocked_timestamp_data.is_ok() {
-            // TODO: handle other errors that are not 404.
-            let template = AlreadyUnlockedTemplate {};
-            return HttpResponse::UnprocessableEntity().body(template.render().unwrap());
-        }
-
-        app_data
-            .seed_repository
-            .move_prefix(seed_name, "locked", "public")
-            .await
-            .unwrap();
-        let timestamp = match SystemTime::now().duration_since(SystemTime::UNIX_EPOCH) {
-            Ok(n) => n.as_millis() as usize,
-            Err(_) => panic!("SystemTime before UNIX EPOCH!"),
-        };
-        let unlock_time_str = format!("{}", timestamp);
-        app_data
-            .seed_repository
-            .put_file(
-                seed_name,
-                "unlocked_timestamp.txt".to_string(),
-                unlock_time_str.into_bytes(),
-            )
-            .await
-            .unwrap();
-    } else {
-        let template = InvalidTokenTemplate {};
-        return HttpResponse::Forbidden().body(template.render().unwrap());
-    }
-    HttpResponse::Found()
-        .insert_header((header::LOCATION, format!("/seed/{}/", info.0)))
-        .finish()
-}
-
-fn get_spin_lock_buttons(req: &CustomizeRequest) -> Vec<ControllerButton> {
-    let mut spin_lock_buttons = vec![];
-    let setting_button_mapping = vec![
-        (&req.spin_lock_left, ControllerButton::Left),
-        (&req.spin_lock_right, ControllerButton::Right),
-        (&req.spin_lock_up, ControllerButton::Up),
-        (&req.spin_lock_down, ControllerButton::Down),
-        (&req.spin_lock_a, ControllerButton::A),
-        (&req.spin_lock_b, ControllerButton::B),
-        (&req.spin_lock_x, ControllerButton::X),
-        (&req.spin_lock_y, ControllerButton::Y),
-        (&req.spin_lock_l, ControllerButton::L),
-        (&req.spin_lock_r, ControllerButton::R),
-        (&req.spin_lock_select, ControllerButton::Select),
-        (&req.spin_lock_start, ControllerButton::Start),
-    ];
-
-    for (setting, button) in setting_button_mapping {
-        if let Some(x) = setting {
-            if x.0 == "on" {
-                spin_lock_buttons.push(button);
-            }
-        }
-    }
-    spin_lock_buttons
-}
-
-fn get_quick_reload_buttons(req: &CustomizeRequest) -> Vec<ControllerButton> {
-    let mut quick_reload_buttons = vec![];
-    let setting_button_mapping = vec![
-        (&req.quick_reload_left, ControllerButton::Left),
-        (&req.quick_reload_right, ControllerButton::Right),
-        (&req.quick_reload_up, ControllerButton::Up),
-        (&req.quick_reload_down, ControllerButton::Down),
-        (&req.quick_reload_a, ControllerButton::A),
-        (&req.quick_reload_b, ControllerButton::B),
-        (&req.quick_reload_x, ControllerButton::X),
-        (&req.quick_reload_y, ControllerButton::Y),
-        (&req.quick_reload_l, ControllerButton::L),
-        (&req.quick_reload_r, ControllerButton::R),
-        (&req.quick_reload_select, ControllerButton::Select),
-        (&req.quick_reload_start, ControllerButton::Start),
-    ];
-
-    for (setting, button) in setting_button_mapping {
-        if let Some(x) = setting {
-            if x.0 == "on" {
-                quick_reload_buttons.push(button);
-            }
-        }
-    }
-    quick_reload_buttons
-}
-
-#[post("/seed/{name}/customize")]
-async fn customize_seed(
-    req: MultipartForm<CustomizeRequest>,
-    info: web::Path<(String,)>,
-    app_data: web::Data<AppData>,
-) -> impl Responder {
-    let seed_name = &info.0;
-    let patch_ips = app_data
-        .seed_repository
-        .get_file(seed_name, "patch.ips")
-        .await
-        .unwrap();
-    let orig_rom = Rom::new(req.rom.data.to_vec());
-    let mut rom = orig_rom.clone();
-
-    let seed_data_str: String = String::from_utf8(
-        app_data
-            .seed_repository
-            .get_file(seed_name, "seed_data.json")
-            .await
-            .unwrap(),
-    )
-    .unwrap();
-    let seed_data = json::parse(&seed_data_str).unwrap();
-
-    let map_data_bytes = app_data
-        .seed_repository
-        .get_file(seed_name, "map.json")
-        .await
-        .unwrap_or(vec![]);
-    let map: Option<Map> = if map_data_bytes.len() == 0 {
-        None
-    } else {
-        Some(serde_json::from_slice(&map_data_bytes).unwrap())
-    };
-
-    let ultra_low_qol = seed_data["ultra_low_qol"].as_bool().unwrap_or(false);
-
-    let rom_digest = crypto_hash::hex_digest(crypto_hash::Algorithm::SHA256, &rom.data);
-    info!("Rom digest: {rom_digest}");
-    if rom_digest != "12b77c4bc9c1832cee8881244659065ee1d84c70c3d29e6eaf92e6798cc2ca72" {
-        return HttpResponse::BadRequest().body(InvalidRomTemplate {}.render().unwrap());
-    }
-
-    let settings = CustomizeSettings {
-        samus_sprite: if ultra_low_qol
-            && req.samus_sprite.0 == "samus"
-            && req.vanilla_screw_attack_animation.0
-        {
-            None
-        } else {
-            Some(req.samus_sprite.0.clone())
-        },
-        etank_color: Some((
-            u8::from_str_radix(&req.etank_color.0[0..2], 16).unwrap() / 8,
-            u8::from_str_radix(&req.etank_color.0[2..4], 16).unwrap() / 8,
-            u8::from_str_radix(&req.etank_color.0[4..6], 16).unwrap() / 8,
-        )),
-        reserve_hud_style: req.reserve_hud_style.0,
-        vanilla_screw_attack_animation: req.vanilla_screw_attack_animation.0,
-        palette_theme: if req.room_palettes.0 == "area-themed" {
-            PaletteTheme::AreaThemed
-        } else {
-            PaletteTheme::Vanilla
-        },
-        tile_theme: if req.tile_theme.0 == "none" {
-            TileTheme::Vanilla
-        } else if req.tile_theme.0 == "scrambled" {
-            TileTheme::Scrambled
-        } else if req.tile_theme.0 == "area_themed" {
-            TileTheme::AreaThemed
-        } else {
-            TileTheme::Constant(req.tile_theme.0.to_string())
-        },
-        door_theme: match req.door_theme.0.as_str() {
-            "vanilla" => DoorTheme::Vanilla,
-            "alternate" => DoorTheme::Alternate,
-            _ => panic!(
-                "Unexpected door_theme option: {}",
-                req.door_theme.0.as_str()
-            ),
-        },
-        music: match req.music.0.as_str() {
-            "vanilla" => MusicSettings::Vanilla,
-            "area" => MusicSettings::AreaThemed,
-            "disabled" => MusicSettings::Disabled,
-            _ => panic!("Unexpected music option: {}", req.music.0.as_str()),
-        },
-        disable_beeping: req.disable_beeping.0,
-        shaking: match req.shaking.0.as_str() {
-            "Vanilla" => ShakingSetting::Vanilla,
-            "Reduced" => ShakingSetting::Reduced,
-            "Disabled" => ShakingSetting::Disabled,
-            _ => panic!("Unexpected shaking option: {}", req.shaking.0.as_str()),
-        },
-        controller_config: ControllerConfig {
-            shot: parse_controller_button(&req.control_shot.0).unwrap(),
-            jump: parse_controller_button(&req.control_jump.0).unwrap(),
-            dash: parse_controller_button(&req.control_dash.0).unwrap(),
-            item_select: parse_controller_button(&req.control_item_select.0).unwrap(),
-            item_cancel: parse_controller_button(&req.control_item_cancel.0).unwrap(),
-            angle_up: parse_controller_button(&req.control_angle_up.0).unwrap(),
-            angle_down: parse_controller_button(&req.control_angle_down.0).unwrap(),
-            spin_lock_buttons: get_spin_lock_buttons(&req),
-            quick_reload_buttons: get_quick_reload_buttons(&req),
-            moonwalk: req.moonwalk.0,
-        },
-    };
-    info!("CustomizeSettings: {:?}", settings);
-    match customize_rom(
-        &mut rom,
-        &orig_rom,
-        &patch_ips,
-        &map,
-        &settings,
-        &app_data.game_data,
-        &app_data.samus_sprite_categories,
-        &app_data.mosaic_themes,
-    ) {
-        Ok(()) => {}
-        Err(err) => {
-            return HttpResponse::InternalServerError()
-                .body(format!("Error customizing ROM: {:?}", err))
-        }
-    }
-
-    HttpResponse::Ok()
-        .content_type("application/octet-stream")
-        .insert_header(ContentDisposition {
-            disposition: DispositionType::Attachment,
-            parameters: vec![DispositionParam::Filename(
-                "map-rando-".to_string() + seed_name + ".sfc",
-            )],
-        })
-        .body(rom.data)
-}
-
-#[get("/seed/{name}/data/{filename:.*}")]
-async fn get_seed_file(
-    info: web::Path<(String, String)>,
-    app_data: web::Data<AppData>,
-) -> impl Responder {
-    let seed_name = &info.0;
-    let filename = &info.1;
-    println!("get_seed_file {}", filename);
-
-    let data_result: Result<Vec<u8>> = if filename.starts_with("visualizer/")
-        && app_data.static_visualizer
-    {
-        let path = Path::new(VISUALIZER_PATH).join(filename.strip_prefix("visualizer/").unwrap());
-        std::fs::read(&path)
-            .map_err(anyhow::Error::from)
-            .with_context(|| format!("Error reading static file: {}", path.display()))
-    } else {
-        app_data
-            .seed_repository
-            .get_file(seed_name, &("public/".to_string() + filename))
-            .await
-    };
-
-    match data_result {
-        Ok(data) => {
-            let ext = Path::new(filename)
-                .extension()
-                .map(|x| x.to_str().unwrap())
-                .unwrap_or("bin");
-            let mime = actix_files::file_extension_to_mime(ext);
-            HttpResponse::Ok().content_type(mime).body(data)
-        }
-        // TODO: Use more refined error handling instead of always returning 404:
-        Err(err) => {
-            error!("{}", err.to_string());
-            HttpResponse::NotFound().body(FileNotFoundTemplate {}.render().unwrap())
-        }
-    }
-}
-
-fn format_http_headers(req: &HttpRequest) -> serde_json::Map<String, serde_json::Value> {
-    let map: serde_json::Map<String, serde_json::Value> = req
-        .headers()
-        .into_iter()
-        .map(|(name, value)| {
-            (
-                name.to_string(),
-                serde_json::Value::String(value.to_str().unwrap_or("").to_string()),
-            )
-        })
-        .collect();
-    map
-}
-
-fn get_random_seed() -> usize {
-    (rand::rngs::StdRng::from_entropy().next_u64() & 0xFFFFFFFF) as usize
-}
-
-// A simplified measure of "equivalence" in difficulty for the purposes of Forced mode logic, to reduce
-// the amount of tiers the randomizer has to consider:
-fn is_equivalent_difficulty(a: &DifficultyConfig, b: &DifficultyConfig) -> bool {
-    return a.tech == b.tech && a.notable_strats == b.notable_strats;
-}
-
-// Computes the intersection of the selected difficulty with each preset. This
-// gives a set of difficulty tiers below the selected difficulty. These are
-// used in "forced mode" to try to identify locations at which to place
-// key items which are reachable using the selected difficulty but not at
-// lower difficulties.
-fn get_difficulty_tiers(
-    difficulty: &DifficultyConfig,
-    app_data: &AppData,
-) -> Vec<DifficultyConfig> {
-    let presets = &app_data.preset_data;
-    let mut out: Vec<DifficultyConfig> = vec![];
-    let tech_set: HashSet<String> = difficulty.tech.iter().cloned().collect();
-    let strat_set: HashSet<String> = difficulty.notable_strats.iter().cloned().collect();
-
-    out.push(difficulty.clone());
-    out.last_mut().unwrap().tech.sort();
-    out.last_mut().unwrap().notable_strats.sort();
-    for preset_data in presets.iter().rev() {
-        let preset = &preset_data.preset;
-        let mut tech_vec: Vec<String> = Vec::new();
-        for (tech, enabled) in &preset_data.tech_setting {
-            if *enabled && tech_set.contains(tech) {
-                tech_vec.push(tech.clone());
-            }
-        }
-        tech_vec.sort();
-
-        let mut strat_vec: Vec<String> = vec![];
-        for (strat, enabled) in &preset_data.notable_strat_setting {
-            if *enabled && strat_set.contains(strat) {
-                strat_vec.push(strat.clone());
-            }
-        }
-        strat_vec.sort();
-
-        // TODO: move some fields out of here so we don't have clone as much irrelevant stuff:
-        let new_difficulty = DifficultyConfig {
-            name: Some(preset.name.clone()),
-            tech: tech_vec,
-            notable_strats: strat_vec,
-            shine_charge_tiles: f32::max(
-                difficulty.shine_charge_tiles,
-                preset.shinespark_tiles as f32,
-            ),
-            heated_shine_charge_tiles: f32::max(
-                difficulty.heated_shine_charge_tiles,
-                preset.heated_shinespark_tiles as f32,
-            ),
-            speed_ball_tiles: f32::max(difficulty.speed_ball_tiles, preset.speed_ball_tiles as f32),
-            shinecharge_leniency_frames: Capacity::max(
-                difficulty.shinecharge_leniency_frames,
-                preset.shinecharge_leniency_frames as Capacity,
-            ),
-            progression_rate: difficulty.progression_rate,
-            random_tank: difficulty.random_tank,
-            spazer_before_plasma: difficulty.spazer_before_plasma,
-            stop_item_placement_early: difficulty.stop_item_placement_early,
-            item_placement_style: difficulty.item_placement_style,
-            item_priority_strength: difficulty.item_priority_strength,
-            item_priorities: difficulty.item_priorities.clone(),
-            item_pool: difficulty.item_pool.clone(),
-            starting_items: difficulty.starting_items.clone(),
-            semi_filler_items: difficulty.semi_filler_items.clone(),
-            filler_items: difficulty.filler_items.clone(),
-            early_filler_items: difficulty.early_filler_items.clone(),
-            resource_multiplier: f32::max(
-                difficulty.resource_multiplier,
-                preset.resource_multiplier,
-            ),
-            gate_glitch_leniency: Capacity::max(
-                difficulty.gate_glitch_leniency,
-                preset.gate_glitch_leniency as Capacity,
-            ),
-            door_stuck_leniency: Capacity::max(
-                difficulty.door_stuck_leniency,
-                preset.door_stuck_leniency as Capacity,
-            ),
-            escape_timer_multiplier: difficulty.escape_timer_multiplier,
-            start_location_mode: difficulty.start_location_mode,
-            save_animals: difficulty.save_animals,
-            phantoon_proficiency: f32::min(
-                difficulty.phantoon_proficiency,
-                preset.phantoon_proficiency,
-            ),
-            draygon_proficiency: f32::min(
-                difficulty.draygon_proficiency,
-                preset.draygon_proficiency,
-            ),
-            ridley_proficiency: f32::min(difficulty.ridley_proficiency, preset.ridley_proficiency),
-            botwoon_proficiency: f32::min(
-                difficulty.botwoon_proficiency,
-                preset.botwoon_proficiency,
-            ),
-            mother_brain_proficiency: f32::min(
-                difficulty.mother_brain_proficiency,
-                preset.mother_brain_proficiency,
-            ),
-            // Quality-of-life options:
-            supers_double: difficulty.supers_double,
-            mother_brain_fight: difficulty.mother_brain_fight,
-            escape_enemies_cleared: difficulty.escape_enemies_cleared,
-            escape_refill: difficulty.escape_refill,
-            escape_movement_items: difficulty.escape_movement_items,
-            mark_map_stations: difficulty.mark_map_stations,
-            room_outline_revealed: difficulty.room_outline_revealed,
-            transition_letters: difficulty.transition_letters,
-            door_locks_size: difficulty.door_locks_size,
-            item_markers: difficulty.item_markers,
-            item_dot_change: difficulty.item_dot_change,
-            all_items_spawn: difficulty.all_items_spawn,
-            acid_chozo: difficulty.acid_chozo,
-            buffed_drops: difficulty.buffed_drops,
-            fast_elevators: difficulty.fast_elevators,
-            fast_doors: difficulty.fast_doors,
-            fast_pause_menu: difficulty.fast_pause_menu,
-            respin: difficulty.respin,
-            infinite_space_jump: difficulty.infinite_space_jump,
-            momentum_conservation: difficulty.momentum_conservation,
-            objectives: difficulty.objectives.clone(),
-            doors_mode: difficulty.doors_mode,
-            early_save: difficulty.early_save,
-            area_assignment: difficulty.area_assignment,
-            wall_jump: difficulty.wall_jump,
-            etank_refill: difficulty.etank_refill,
-            maps_revealed: difficulty.maps_revealed,
-            map_station_reveal: difficulty.map_station_reveal,
-            energy_free_shinesparks: difficulty.energy_free_shinesparks,
-            vanilla_map: difficulty.vanilla_map,
-            ultra_low_qol: difficulty.ultra_low_qol,
-            skill_assumptions_preset: difficulty.skill_assumptions_preset.clone(),
-            item_progression_preset: difficulty.item_progression_preset.clone(),
-            quality_of_life_preset: difficulty.quality_of_life_preset.clone(),
-            debug_options: difficulty.debug_options.clone(),
-        };
-        if is_equivalent_difficulty(&new_difficulty, out.last().as_ref().unwrap()) {
-            out.pop();
-        }
-        out.push(new_difficulty);
-    }
-    out
-}
-
-fn get_item_priorities(item_priority_json: serde_json::Value) -> Vec<ItemPriorityGroup> {
-    let mut priorities: IndexedVec<String> = IndexedVec::default();
-    priorities.add("Early");
-    priorities.add("Default");
-    priorities.add("Late");
-
-    let mut out: Vec<ItemPriorityGroup> = Vec::new();
-    for priority in &priorities.keys {
-        out.push(ItemPriorityGroup {
-            name: priority.clone(),
-            items: vec![],
-        });
-    }
-    for (k, v) in item_priority_json.as_object().unwrap() {
-        let i = priorities.index_by_key[v.as_str().unwrap()];
-        out[i].items.push(k.to_string());
-    }
-    out
-}
-
-#[post("/randomize")]
-async fn randomize(
-    req: MultipartForm<RandomizeRequest>,
-    http_req: HttpRequest,
-    app_data: web::Data<AppData>,
-) -> impl Responder {
-    let rom = Rom::new(req.rom.data.to_vec());
-
-    if rom.data.len() == 0 {
-        return HttpResponse::BadRequest().body(MissingInputRomTemplate {}.render().unwrap());
-    }
-
-    let rom_digest = crypto_hash::hex_digest(crypto_hash::Algorithm::SHA256, &rom.data);
-    info!("Rom digest: {rom_digest}");
-    if rom_digest != "12b77c4bc9c1832cee8881244659065ee1d84c70c3d29e6eaf92e6798cc2ca72" {
-        return HttpResponse::BadRequest().body(InvalidRomTemplate {}.render().unwrap());
-    }
-
-    let race_mode = req.race_mode.0 == "Yes";
-    let random_seed = if &req.random_seed.0 == "" || race_mode {
-        get_random_seed()
-    } else {
-        match req.random_seed.0.parse::<usize>() {
-            Ok(x) => x,
-            Err(_) => {
-                return HttpResponse::BadRequest().body("Invalid random seed");
-            }
-        }
-    };
-    let display_seed = if race_mode {
-        get_random_seed()
-    } else {
-        random_seed
-    };
-
-    if req.ridley_proficiency.0 < 0.0 || req.ridley_proficiency.0 > 1.0 {
-        return HttpResponse::BadRequest().body("Invalid Ridley proficiency");
-    }
-
-    let tech_json: serde_json::Value = serde_json::from_str(&req.tech_json).unwrap();
-    let mut tech_vec: Vec<String> = app_data.implicit_tech.iter().cloned().collect();
-    let walljump_tech = "canWalljump";
-    assert!(tech_json.as_object().unwrap().contains_key(walljump_tech));
-    for (tech, is_enabled) in tech_json.as_object().unwrap().iter() {
-        if tech == walljump_tech && req.wall_jump.0 == "Disabled" {
-            continue;
-        }
-        if is_enabled.as_bool().unwrap() {
-            tech_vec.push(tech.to_string());
-        }
-    }
-
-    let vanilla_map = req.map_layout.0 == "Vanilla";
-    if vanilla_map {
-        tech_vec.push("canEscapeMorphLocation".to_string());
-    }
-
-    let strat_json: serde_json::Value = serde_json::from_str(&req.strat_json).unwrap();
-    let mut strat_vec: Vec<String> = vec![];
-    for (strat, is_enabled) in strat_json.as_object().unwrap().iter() {
-        if is_enabled.as_bool().unwrap() {
-            strat_vec.push(strat.to_string());
-        }
-    }
-
-    let starting_item_json: serde_json::Value =
-        serde_json::from_str(&req.starting_item_json).unwrap();
-    info!("starting_item_json: {:?}", starting_item_json);
-    let mut starting_items: Vec<(Item, usize)> = vec![];
-    for (k, cnt_json) in starting_item_json.as_object().unwrap().iter() {
-        let cnt_str = cnt_json.as_str().unwrap();
-        let cnt = if cnt_str == "Yes" {
-            1
-        } else if cnt_str == "No" {
-            0
-        } else {
-            usize::from_str_radix(cnt_str, 10).unwrap()
-        };
-        if cnt > 0 {
-            starting_items.push((
-                Item::try_from(app_data.game_data.item_isv.index_by_key[k]).unwrap(),
-                cnt,
-            ));
-        }
-    }
-    info!("Starting items: {:?}", starting_items);
-
-    let item_pool_json: serde_json::Value = serde_json::from_str(&req.item_pool_json).unwrap();
-    info!("item_pool_json: {:?}", item_pool_json);
-    let mut item_pool: Vec<(Item, usize)> = vec![];
-    for (k, cnt_json) in item_pool_json.as_object().unwrap().iter() {
-        let cnt_str = cnt_json.as_str().unwrap();
-        let cnt = usize::from_str_radix(cnt_str, 10).unwrap();
-        item_pool.push((
-            Item::try_from(app_data.game_data.item_isv.index_by_key[k]).unwrap(),
-            cnt,
-        ));
-    }
-    info!("Item pool: {:?}", item_pool);
-
-    let item_priority_json: serde_json::Value =
-        serde_json::from_str(&req.item_priority_json.0).unwrap();
-
-    let filler_items_json: serde_json::Value =
-        serde_json::from_str(&req.filler_items_json.0).unwrap();
-    let semi_filler_items: Vec<Item> = filler_items_json
-        .as_object()
-        .unwrap()
-        .iter()
-        .filter(|(_k, v)| v.as_str().unwrap() == "Semi")
-        .map(|(k, _v)| Item::try_from(app_data.game_data.item_isv.index_by_key[k]).unwrap())
-        .collect();
-    let mut filler_items = vec![Item::Missile, Item::Nothing];
-    filler_items.extend(
-        filler_items_json
-            .as_object()
-            .unwrap()
-            .iter()
-            .filter(|(_k, v)| v.as_str().unwrap() == "Yes" || v.as_str().unwrap() == "Early")
-            .map(|(k, _v)| Item::try_from(app_data.game_data.item_isv.index_by_key[k]).unwrap()),
-    );
-    let early_filler_items: Vec<Item> = filler_items_json
-        .as_object()
-        .unwrap()
-        .iter()
-        .filter(|(_k, v)| v.as_str().unwrap() == "Early")
-        .map(|(k, _v)| Item::try_from(app_data.game_data.item_isv.index_by_key[k]).unwrap())
-        .collect();
-
-    info!("Filler items: {:?}", filler_items);
-    info!("Semi-filler items: {:?}", semi_filler_items);
-    info!("Early filler items: {:?}", early_filler_items);
-
-    let mut rng_seed = [0u8; 32];
-    rng_seed[..8].copy_from_slice(&random_seed.to_le_bytes());
-    let mut rng = rand::rngs::StdRng::from_seed(rng_seed);
-
-    let difficulty = DifficultyConfig {
-        name: None,
-        tech: tech_vec,
-        notable_strats: strat_vec,
-        shine_charge_tiles: req.shinespark_tiles.0,
-        heated_shine_charge_tiles: req.heated_shinespark_tiles.0,
-        speed_ball_tiles: req.speed_ball_tiles.0,
-        shinecharge_leniency_frames: req.shinecharge_leniency_frames.0,
-        progression_rate: match req.progression_rate.0.as_str() {
-            "Slow" => maprando::randomize::ProgressionRate::Slow,
-            "Uniform" => maprando::randomize::ProgressionRate::Uniform,
-            "Fast" => maprando::randomize::ProgressionRate::Fast,
-            _ => panic!(
-                "Unrecognized progression rate {}",
-                req.progression_rate.0.as_str()
-            ),
-        },
-        item_priority_strength: match req.item_priority_strength.0.as_str() {
-            "Moderate" => ItemPriorityStrength::Moderate,
-            "Heavy" => ItemPriorityStrength::Heavy,
-            _ => panic!(
-                "Unrecognized item priority strength {}",
-                req.item_priority_strength.0,
-            ),
-        },
-        random_tank: match req.random_tank.0.as_str() {
-            "No" => false,
-            "Yes" => true,
-            _ => panic!("Unrecognized random_tank {}", req.random_tank.0.as_str()),
-        },
-        spazer_before_plasma: match req.spazer_before_plasma.0.as_str() {
-            "No" => false,
-            "Yes" => true,
-            _ => panic!(
-                "Unrecognized spazer_before_plasma {}",
-                req.spazer_before_plasma.0.as_str()
-            ),
-        },
-        stop_item_placement_early: match req.stop_item_placement_early.0.as_str() {
-            "No" => false,
-            "Yes" => true,
-            _ => panic!(
-                "Unrecognized stop_item_placement_early {}",
-                req.stop_item_placement_early.0.as_str()
-            ),
-        },
-        item_pool,
-        starting_items,
-        filler_items,
-        semi_filler_items,
-        early_filler_items,
-        item_placement_style: match req.item_placement_style.0.as_str() {
-            "Neutral" => maprando::randomize::ItemPlacementStyle::Neutral,
-            "Forced" => maprando::randomize::ItemPlacementStyle::Forced,
-            _ => panic!(
-                "Unrecognized item placement style {}",
-                req.item_placement_style.0.as_str()
-            ),
-        },
-        item_priorities: get_item_priorities(item_priority_json),
-        resource_multiplier: req.resource_multiplier.0,
-        escape_timer_multiplier: req.escape_timer_multiplier.0,
-        gate_glitch_leniency: req.gate_glitch_leniency.0,
-        door_stuck_leniency: req.door_stuck_leniency.0,
-        start_location_mode: match req.start_location.0.as_str() {
-            "Ship" => maprando::randomize::StartLocationMode::Ship,
-            "Random" => maprando::randomize::StartLocationMode::Random,
-            "Escape" => maprando::randomize::StartLocationMode::Escape,
-            _ => panic!("Unrecognized start_location: {}", req.start_location.0),
-        },
-        save_animals: match req.save_animals.0.as_str() {
-            "No" => SaveAnimals::No,
-            "Maybe" => SaveAnimals::Maybe,
-            "Yes" => SaveAnimals::Yes,
-            _ => panic!(
-                "Unrecognized save_animals options {}",
-                req.save_animals.0.as_str()
-            ),
-        },
-        phantoon_proficiency: req.phantoon_proficiency.0,
-        draygon_proficiency: req.draygon_proficiency.0,
-        ridley_proficiency: req.ridley_proficiency.0,
-        botwoon_proficiency: req.botwoon_proficiency.0,
-        mother_brain_proficiency: req.mother_brain_proficiency.0,
-        supers_double: req.supers_double.0,
-        mother_brain_fight: match req.mother_brain_fight.0.as_str() {
-            "Vanilla" => MotherBrainFight::Vanilla,
-            "Short" => MotherBrainFight::Short,
-            "Skip" => MotherBrainFight::Skip,
-            _ => panic!(
-                "Unrecognized mother_brain_fight: {}",
-                req.mother_brain_fight.0
-            ),
-        },
-        escape_enemies_cleared: req.escape_enemies_cleared.0,
-        escape_refill: req.escape_refill.0,
-        escape_movement_items: req.escape_movement_items.0,
-        mark_map_stations: req.mark_map_stations.0,
-        room_outline_revealed: req.room_outline_revealed.0,
-        transition_letters: req.transition_letters.0,
-        door_locks_size: match req.door_locks_size.0.as_str() {
-            "small" => DoorLocksSize::Small,
-            "large" => DoorLocksSize::Large,
-            _ => panic!("Unrecognized door_locks_size: {}", req.door_locks_size.0),
-        },
-        item_markers: match req.item_markers.0.as_str() {
-            "Simple" => ItemMarkers::Simple,
-            "Majors" => ItemMarkers::Majors,
-            "Uniques" => ItemMarkers::Uniques,
-            "3-Tiered" => ItemMarkers::ThreeTiered,
-            _ => panic!("Unrecognized item_markers: {}", req.item_markers.0),
-        },
-        item_dot_change: match req.item_dot_change.0.as_str() {
-            "Fade" => ItemDotChange::Fade,
-            "Disappear" => ItemDotChange::Disappear,
-            _ => panic!("Unrecognized item_dot_change: {}", req.item_dot_change.0),
-        },
-        all_items_spawn: req.all_items_spawn.0,
-        acid_chozo: req.acid_chozo.0,
-        buffed_drops: req.buffed_drops.0,
-        fast_elevators: req.fast_elevators.0,
-        fast_doors: req.fast_doors.0,
-        fast_pause_menu: req.fast_pause_menu.0,
-        respin: req.respin.0,
-        infinite_space_jump: req.infinite_space_jump.0,
-        momentum_conservation: req.momentum_conservation.0,
-        objectives: {
-            use Objective::*;
-            match req.objectives.0.as_str() {
-                "None" => vec![],
-                "Bosses" => vec![Kraid, Phantoon, Draygon, Ridley],
-                "Minibosses" => vec![SporeSpawn, Crocomire, Botwoon, GoldenTorizo],
-                "Metroids" => vec![MetroidRoom1, MetroidRoom2, MetroidRoom3, MetroidRoom4],
-                "Chozos" => vec![BombTorizo, BowlingStatue, AcidChozoStatue, GoldenTorizo],
-                "Pirates" => vec![PitRoom, BabyKraidRoom, PlasmaRoom, MetalPiratesRoom],
-                "Random" => {
-                    rand::seq::SliceRandom::choose_multiple(Objective::get_all(), &mut rng, 4)
-                        .copied()
-                        .collect()
-                }
-                _ => panic!("Unrecognized objectives: {}", req.objectives.0),
-            }
-        },
-        doors_mode: match req.doors.0.as_str() {
-            "Blue" => DoorsMode::Blue,
-            "Ammo" => DoorsMode::Ammo,
-            "Beam" => DoorsMode::Beam,
-            _ => panic!("Unrecognized doors mode: {}", req.doors.0),
-        },
-        early_save: req.early_save.0,
-        area_assignment: match req.area_assignment.0.as_str() {
-            "Standard" => AreaAssignment::Standard,
-            "Random" => AreaAssignment::Random,
-            _ => panic!("Unrecognized ship area option: {}", req.area_assignment.0),
-        },
-        wall_jump: if req.start_location.0.as_str() == "Escape" {
-            maprando::randomize::WallJump::Vanilla
-        } else {
-            match req.wall_jump.0.as_str() {
-                "Vanilla" => maprando::randomize::WallJump::Vanilla,
-                "Collectible" => maprando::randomize::WallJump::Collectible,
-                _ => panic!(
-                    "Unrecognized wall_jump setting {}",
-                    req.wall_jump.0.as_str()
-                ),
-            }
-        },
-        etank_refill: match req.etank_refill.0.as_str() {
-            "Disabled" => maprando::randomize::EtankRefill::Disabled,
-            "Vanilla" => maprando::randomize::EtankRefill::Vanilla,
-            "Full" => maprando::randomize::EtankRefill::Full,
-            _ => panic!(
-                "Unrecognized etank_refill setting {}",
-                req.etank_refill.0.as_str()
-            ),
-        },
-        maps_revealed: match req.maps_revealed.0.as_str() {
-            "No" => maprando::randomize::MapsRevealed::No,
-            "Partial" => maprando::randomize::MapsRevealed::Partial,
-            "Full" => maprando::randomize::MapsRevealed::Full,
-            _ => panic!(
-                "Unrecognized maps_revealed setting {}",
-                req.maps_revealed.0.as_str()
-            ),
-        },
-        map_station_reveal: match req.map_station_reveal.0.as_str() {
-            "Partial" => maprando::randomize::MapStationReveal::Partial,
-            "Full" => maprando::randomize::MapStationReveal::Full,
-            _ => panic!(
-                "Unrecognized maps_station_reveal setting {}",
-                req.maps_revealed.0.as_str()
-            ),
-        },
-        vanilla_map,
-        energy_free_shinesparks: req.energy_free_shinesparks.0,
-        ultra_low_qol: req.ultra_low_qol.0,
-        skill_assumptions_preset: req.preset.as_ref().map(|x| x.0.clone()),
-        item_progression_preset: req.item_progression_preset.as_ref().map(|x| x.0.clone()),
-        quality_of_life_preset: req.quality_of_life_preset.as_ref().map(|x| x.0.clone()),
-        debug_options: if app_data.debug {
-            Some(DebugOptions {
-                new_game_extra: true,
-                extended_spoiler: true,
-            })
-        } else {
-            None
-        },
-    };
-    let difficulty_tiers = if difficulty.item_placement_style == ItemPlacementStyle::Forced {
-        get_difficulty_tiers(&difficulty, &app_data)
-    } else {
-        vec![difficulty.clone()]
-    };
-
-    let filtered_base_links =
-        filter_links(&app_data.game_data.links, &app_data.game_data, &difficulty);
-    let filtered_base_links_data = LinksDataGroup::new(
-        filtered_base_links,
-        app_data.game_data.vertex_isv.keys.len(),
-        0,
-    );
-    let map_layout = req.map_layout.0.clone();
-    let max_attempts = 2000;
-    let max_attempts_per_map = if difficulty.start_location_mode == StartLocationMode::Random {
-        10
-    } else {
-        1
-    };
-    let max_map_attempts = max_attempts / max_attempts_per_map;
-    info!(
-        "Random seed={random_seed}, max_attempts_per_map={max_attempts_per_map}, max_map_attempts={max_map_attempts}, difficulty={:?}",
-        difficulty_tiers[0]
-    );
-
-    struct AttemptOutput {
-        map_seed: usize,
-        door_randomization_seed: usize,
-        item_placement_seed: usize,
-        randomization: Randomization,
-        output_rom: Rom,
-    }
-
-    let time_start_attempts = Instant::now();
-    let mut attempt_num = 0;
-    let mut output_opt: Option<AttemptOutput> = None;
-    'attempts: for _ in 0..max_map_attempts {
-        let map_seed = (rng.next_u64() & 0xFFFFFFFF) as usize;
-        let door_randomization_seed = (rng.next_u64() & 0xFFFFFFFF) as usize;
-
-        if !app_data.map_repositories.contains_key(&map_layout) {
-            // TODO: it doesn't make sense to panic on things like this.
-            panic!("Unrecognized map layout option: {}", map_layout);
-        }
-        let mut map = app_data.map_repositories[&map_layout]
-            .get_map(attempt_num, map_seed, &app_data.game_data)
-            .unwrap();
-        if difficulty.area_assignment == AreaAssignment::Random {
-            randomize_map_areas(&mut map, map_seed);
-        }
-        let locked_door_data = randomize_doors(
-            &app_data.game_data,
-            &map,
-            &difficulty_tiers[0],
-            door_randomization_seed,
-        );
-        let randomizer = Randomizer::new(
-            &map,
-            &locked_door_data,
-            &difficulty_tiers,
-            &app_data.game_data,
-            &filtered_base_links_data,
-        );
-        for _ in 0..max_attempts_per_map {
-            let item_placement_seed = (rng.next_u64() & 0xFFFFFFFF) as usize;
-            attempt_num += 1;
-
-            info!("Attempt {attempt_num}/{max_attempts}: Map seed={map_seed}, door randomization seed={door_randomization_seed}, item placement seed={item_placement_seed}");
-            let randomization_result =
-                randomizer.randomize(attempt_num, item_placement_seed, display_seed);
-            let randomization = match randomization_result {
-                Ok(x) => x,
-                Err(e) => {
-                    info!(
-                        "Attempt {attempt_num}/{max_attempts}: Randomization failed: {}",
-                        e
-                    );
-                    continue;
-                }
-            };
-            let output_rom_result = make_rom(&rom, &randomization, &app_data.game_data);
-            let output_rom = match output_rom_result {
-                Ok(x) => x,
-                Err(e) => {
-                    info!(
-                        "Attempt {attempt_num}/{max_attempts}: Failed to write ROM: {}",
-                        e
-                    );
-                    continue;
-                }
-            };
-            info!(
-                "Successful attempt {attempt_num}/{attempt_num}/{max_attempts}: display_seed={}, random_seed={random_seed}, map_seed={map_seed}, door_randomization_seed={door_randomization_seed}, item_placement_seed={item_placement_seed}",
-                randomization.display_seed,
-            );
-            output_opt = Some(AttemptOutput {
-                map_seed,
-                door_randomization_seed,
-                item_placement_seed,
-                randomization,
-                output_rom,
-            });
-            break 'attempts;
-        }
-    }
-
-    if output_opt.is_none() {
-        return HttpResponse::InternalServerError().body("Failed too many randomization attempts");
-    }
-    let output = output_opt.unwrap();
-
-    info!(
-        "Wall-clock time for attempts: {:?} sec",
-        time_start_attempts.elapsed().as_secs_f32()
-    );
-    let timestamp = match SystemTime::now().duration_since(SystemTime::UNIX_EPOCH) {
-        Ok(n) => n.as_millis() as usize,
-        Err(_) => panic!("SystemTime before UNIX EPOCH!"),
-    };
-    let seed_data = SeedData {
-        version: VERSION,
-        timestamp,
-        peer_addr: http_req
-            .peer_addr()
-            .map(|x| format!("{:?}", x))
-            .unwrap_or(String::new()),
-        http_headers: format_http_headers(&http_req),
-        random_seed: random_seed,
-        map_seed: output.map_seed,
-        door_randomization_seed: output.door_randomization_seed,
-        item_placement_seed: output.item_placement_seed,
-        race_mode,
-        preset: req.preset.as_ref().map(|x| x.0.clone()),
-        item_progression_preset: req.item_progression_preset.as_ref().map(|x| x.0.clone()),
-        difficulty: difficulty_tiers[0].clone(),
-        quality_of_life_preset: req.quality_of_life_preset.as_ref().map(|x| x.0.clone()),
-        supers_double: req.supers_double.0,
-        mother_brain_fight: req.mother_brain_fight.0.clone(),
-        escape_enemies_cleared: req.escape_enemies_cleared.0,
-        escape_refill: req.escape_refill.0,
-        escape_movement_items: req.escape_movement_items.0,
-        mark_map_stations: req.mark_map_stations.0,
-        transition_letters: req.transition_letters.0,
-        item_markers: req.item_markers.0.clone(),
-        item_dot_change: req.item_dot_change.0.clone(),
-        all_items_spawn: req.all_items_spawn.0,
-        acid_chozo: req.acid_chozo.0,
-        buffed_drops: req.buffed_drops.0,
-        fast_elevators: req.fast_elevators.0,
-        fast_doors: req.fast_doors.0,
-        fast_pause_menu: req.fast_pause_menu.0,
-        respin: req.respin.0,
-        infinite_space_jump: req.infinite_space_jump.0,
-        momentum_conservation: req.momentum_conservation.0,
-        objectives: req.objectives.0.clone(),
-        doors: req.doors.0.clone(),
-        start_location_mode: req.start_location.0.clone(),
-        map_layout: req.map_layout.0.clone(),
-        save_animals: req.save_animals.0.clone(),
-        early_save: req.early_save.0,
-        area_assignment: req.area_assignment.0.clone(),
-        wall_jump: req.wall_jump.0.clone(),
-        etank_refill: req.etank_refill.0.clone(),
-        maps_revealed: req.maps_revealed.0.clone(),
-        vanilla_map,
-        ultra_low_qol: req.ultra_low_qol.0,
-    };
-
-    let seed_name = &output.randomization.seed_name;
-    save_seed(
-        seed_name,
-        &seed_data,
-        &req.spoiler_token.0,
-        &rom,
-        &output.output_rom,
-        &output.randomization,
-        &app_data,
-    )
-    .await
-    .unwrap();
-
-    // Redirect to the seed page:
-    HttpResponse::Found()
-        .insert_header((header::LOCATION, format!("seed/{}/", seed_name)))
-        .finish()
-}
 
 fn init_presets(
     presets: Vec<Preset>,
@@ -2212,7 +311,7 @@ async fn main() {
         .format_timestamp_millis()
         .init();
 
-    let app_data = web::Data::new(build_app_data());
+    let app_data = actix_web::web::Data::new(build_app_data());
 
     HttpServer::new(move || {
         App::new()
@@ -2224,20 +323,13 @@ async fn main() {
                     .total_limit(16_000_000),
             )
             .wrap(Logger::default())
-            .service(home)
-            .service(releases)
-            .service(generate)
-            .service(randomize)
-            .service(about)
-            .service(view_seed)
-            .service(get_seed_file)
-            .service(customize_seed)
-            .service(unlock_seed)
-            .service(view_seed_redirect)
-            .service(logic_main)
-            .service(logic_room)
-            .service(logic_strat)
-            .service(logic_tech)
+            .service(home::home)
+            .service(releases::releases)
+            .service(generate::generate)
+            .service(randomize::randomize)
+            .service(about::about)
+            .service(seed::scope())
+            .service(logic::scope())
             .service(actix_files::Files::new(
                 "/static/sm-json-data",
                 "../sm-json-data",
