+++ rust/maprando-web/src/web/logic/logic_main.rs
@@ -1,0 +1,7 @@
+use crate::web::AppData;
+use actix_web::{get, web, HttpResponse, Responder};
+
+#[get("")]
+async fn logic_main(app_data: web::Data<AppData>) -> impl Responder {
+    HttpResponse::Ok().body(app_data.logic_data.index_html.clone())
+}
